dm: gpio: Allow simple access from SPL
Some boards need to enable a UART or other device very early in SPL,
before driver model has been initialised. Add the beginnings of a new GPIO
interface to support this.

This interface is similar to the debug UART. It should only be used when
debugging or for other extreme situations.

Signed-off-by: Simon Glass <sjg@chromium.org>